ajeitando services e criando os controllers com endpoins gets

+++ src/main/java/com/pedrohnf688/api/modelo/Orcamento.java
@@ -27,7 +27,7 @@ public class Orcamento {
 	private Long id;
 	private Date dateCreated;
 	private Integer qtdDiasTrabalhada;
-	private Float custoFinal;
+	private Double custoFinal;
 	private String observacoes;
 	
 	@ManyToOne
